+++ changelog.md
@@ -2,12 +2,25 @@
 
 ## Unreleased
 
+_Nothing yet..._
+
+## 2018-03-15 - Native Panel component
+
+**Added:**
+- react-native-bpk-component-panel: 1.0.0
+  - New native panel component, see https://backpack.github.io/components/native/panels.
+
+- react-native-bpk-component-icon:
+- bpk-component-icon:
+- bpk-mixins:
+- bpk-svgs:
+  - New hotels amenities icons: `hotels--smoking`, `hotels--pets-allowed`, `hotels--jacuzzi` and `hotels--disabled-facilities`.
+
 ## 2018-03-13 - Native Phone Number Input
 
 **Added:**
-- react-native-bpk-component-phone-input:
-  - New `BpkDiallingCodeList` component.
-  - New `BpkPhoneNumberInput` component.
+- react-native-bpk-component-phone-input: 1.0.0
+  - New phone input components, see https://backpack.github.io/components/native/phone-input.
 
 ## 2018-03-12 - Native nudger component and SSR OC fixes